+++ libsnark/prover_reference_functions.cpp
@@ -231,13 +231,215 @@ mnt4753_libsnark::vector_Fr *mnt4753_libsnark::vector_Fr_zeros(size_t length) {
 void mnt4753_libsnark::domain_iFFT(mnt4753_libsnark::evaluation_domain *domain,
                                    mnt4753_libsnark::vector_Fr *a) {
   std::vector<Fr<mnt4753_pp>> &data = *a->data;
+  printf("FFT ===== \n");
   domain->data->iFFT(data);
 }
 
 void mnt4753_libsnark::domain_iFFT_GPU(mnt4753_libsnark::evaluation_domain *domain,
-                                   mnt4753_libsnark::vector_Fr *a) {
+                                   mnt4753_libsnark::vector_Fr *a, Kernel k) {
   std::vector<Fr<mnt4753_pp>> &data = *a->data;
-  domain->data->iFFT(data);
+  k.init(10);
+
+    cl_kernel kernel;                   // compute kernel
+    cl_event event;                     // timing
+    cl_ulong time_start;
+    cl_ulong time_end;
+    int n = 512;
+
+    libff::G1<mnt4753_pp>* data_x = new libff::G1<mnt4753_pp>[1];              // original data set given to device
+    libff::G1<mnt4753_pp>* data_y = new libff::G1<mnt4753_pp>[n];              // original data set given to device
+    libff::G1<mnt4753_pp>* results = new libff::G1<mnt4753_pp>[1];           // results returned from device
+    
+    unsigned int correct;               // number of correct results returned
+
+    cl_mem input_x;                       // device memory used for the input array
+    cl_mem input_y;                       // device memory used for the input array
+    cl_mem ocl_output;                       // device memory used for the input array
+    // Fill our data set with field inputs from param gen
+    //
+    unsigned int count = n;
+    mp_size_t num = 1;
+    Kernel kern;
+    kern.init(n);
+
+    //memcpy(&data_x[0], &h4_1, sizeof(G1<mnt4753_pp>));
+    data_x[0] = libff::G1<mnt4753_pp>::zero();
+    printf("count %u\n", n);
+    data_x[0].print_coordinates();
+
+    for(int i = 0; i < count; i++) {
+      memcpy(&data_y[i], &data_x[0], sizeof(libff::G1<mnt4753_pp>));
+    }
+    
+    
+    // Create the compute kernel in the program we wish to run
+    //
+    kernel = clCreateKernel(kern.program, "multiexp_G1", &kern.err);
+    if (!kernel || kern.err != CL_SUCCESS)
+    {
+        printf("Error: Failed to create compute kernel!\n");
+        exit(1);
+    }
+
+    // Create the input and output arrays in device memory for our calculation
+    //
+    printf("creating buffer\n");
+    input_x = clCreateBuffer(kern.context,  CL_MEM_READ_ONLY,  sizeof(libff::G1<mnt4753_pp>), NULL, NULL);
+    input_y = clCreateBuffer(kern.context,  CL_MEM_READ_ONLY,  sizeof(libff::G1<mnt4753_pp>) * count, NULL, NULL);
+    ocl_output = clCreateBuffer(kern.context, CL_MEM_WRITE_ONLY, sizeof(libff::G1<mnt4753_pp>), NULL, NULL);
+
+    if (!input_x || !ocl_output)
+    {
+        printf("Error: Failed to allocate device memory!\n");
+        exit(1);
+    }
+
+    // Write our data set into the input array in device memory 
+    //
+    auto start = high_resolution_clock::now();
+    kern.err = clEnqueueWriteBuffer(kern.commands, input_x, CL_TRUE, 0, sizeof(libff::G1<mnt4753_pp>), data_x, 0, NULL, NULL);
+    if (kern.err != CL_SUCCESS)
+    {
+        printf("Error: Failed to write to source array!\n");
+        exit(1);
+    }
+    kern.err = clEnqueueWriteBuffer(kern.commands, input_y, CL_TRUE, 0, sizeof(libff::G1<mnt4753_pp>) * count, data_y, 0, NULL, NULL);
+    if (kern.err != CL_SUCCESS)
+    {
+        printf("Error: Failed to write to source array!\n");
+        exit(1);
+    }
+    auto stop = high_resolution_clock::now();
+    auto duration = duration_cast<microseconds>(stop - start); 
+    cout << "Time taken by GPU write function: "
+      << duration.count() << " microseconds" << endl;
+
+    // Set the arguments to our compute kernel
+    //
+    kern.err = 0;
+    kern.err  = clSetKernelArg(kernel, 0, sizeof(cl_mem), &input_x);
+    kern.err  = clSetKernelArg(kernel, 1, sizeof(cl_mem), &input_y);
+    kern.err |= clSetKernelArg(kernel, 2, sizeof(cl_mem), &ocl_output);
+    kern.err |= clSetKernelArg(kernel, 3, sizeof(unsigned int), &count);
+    if (kern.err != CL_SUCCESS)
+    {
+        printf("Error: Failed to set kernel arguments! %d\n", kern.err);
+        exit(1);
+    }
+
+    // Get the maximum work group size for executing the kernel on the device
+    //
+    kern.err = clGetKernelWorkGroupInfo(kernel, kern.devices[0], CL_KERNEL_WORK_GROUP_SIZE, sizeof(kern.local), &kern.local, NULL);
+    if (kern.err != CL_SUCCESS)
+    {
+        printf("Error: Failed to retrieve kernel work group info! %d\n", kern.err);
+        exit(1);
+    }
+
+    printf("Max work size: %u\n", kern.local);
+
+    // Execute the kernel over the entire range of our 1d input data set
+    // using the maximum number of work group items for this device
+    //
+    kern.global = count;
+    //global = 1;
+    printf("queueing kernel\n");
+    kern.err = clEnqueueNDRangeKernel(kern.commands, kernel, 1, NULL, &kern.global, &kern.local, 0, NULL, &event);
+    if (kern.err)
+    {
+        printf("Error: Failed to execute kernel!\n");
+        exit(1);
+    }
+
+    clWaitForEvents(1, &event);
+    clFinish(kern.commands);
+
+    // Time kernel execution time without read/write
+    //
+    clGetEventProfilingInfo(event, CL_PROFILING_COMMAND_START, sizeof(time_start), &time_start, NULL);
+    clGetEventProfilingInfo(event, CL_PROFILING_COMMAND_END, sizeof(time_end), &time_end, NULL);
+
+    double nanoSeconds = time_end-time_start;
+    printf("OpenCl Execution time is: %0.3f milliseconds \n",nanoSeconds / 1000000.0);
+
+    // Read back the results from the device to verify the output
+    //
+    start = high_resolution_clock::now();
+    kern.err = clEnqueueReadBuffer(kern.commands, ocl_output, CL_TRUE, 0, sizeof(libff::G1<mnt4753_pp>), results, 0, NULL, NULL );  
+    if (kern.err != CL_SUCCESS)
+    {
+        printf("Error: Failed to read output array! %d\n", kern.err);
+        exit(1);
+    }
+    stop = high_resolution_clock::now();
+    duration = duration_cast<microseconds>(stop - start); 
+    cout << "Time taken by GPU read function: "
+      << duration.count() << " microseconds" << endl;
+    // Validate our results
+    //
+    printf("Kernel Result \n");
+    results[0].print();
+
+    // results[0].coeff_a.mont_repr.print_hex();
+    // for(int i=0; i<12; i++) {
+    //   //std::cout << "Length of array = " << (sizeof(results[1013].non_residue.mont_repr.data)/sizeof(*results[1013].non_residue.mont_repr.data)) << std::endl;
+    //   cl_uint x;
+    //   cl_uint y;
+    //   x = (cl_uint)((results[0].coeff_a.mont_repr.data[i] & 0xFFFFFFFF00000000LL) >> 32);
+    //   y = (cl_uint)(results[0].coeff_a.mont_repr.data[i] & 0xFFFFFFFFLL);
+    //   gmp_printf("%Mx\n", results[0].coeff_a.mont_repr.data[i]);
+    //   printf("%x\n", x);
+    //   printf("%x\n", y);
+    // }
+
+    results[0].zero().print_coordinates();
+
+    // for(int i=0; i<12; i++) {
+    //   //printf("%x\n", results[1013].c0.mod.data[i]);
+    //   //std::cout << "Length of array = " << (sizeof(results[1013].non_residue.mont_repr.data)/sizeof(*results[1013].non_residue.mont_repr.data)) << std::endl;
+    //   cl_uint x;
+    //   cl_uint y;
+    //   x = (cl_uint)((results[1013].c0.one().mont_repr.data[i] & 0xFFFFFFFF00000000LL) >> 32);
+    //   y = (cl_uint)(results[1013].c0.one().mont_repr.data[i] & 0xFFFFFFFFLL);
+    //   gmp_printf("%Mx\n", results[1013].c0.one().mont_repr.data[i]);
+    //   printf("%x\n", x);
+    //   printf("%x\n", y);
+    // }
+
+    printf("CPU Result\n");
+    //G1<mnt4753_pp> _h4_1 = G1<mnt4753_pp>::zero();
+
+    //for (size_t i = 0; i < n; ++i) { _h4_1 = _h4_1 + g4_1[i]; }
+    // _h4_1 = _h4_1 + g4_1[0];
+    // _h4_1 = _h4_1 + g4_1[1];
+    // _h4_1 = _h4_1 + g4_1[2];
+    // _h4_1 = _h4_1 + g4_1[3];
+    // _h4_1 = _h4_1 + g4_1[4];
+    // _h4_1.print();
+    //  g4_1[1].X().print();
+    correct = 0;
+
+    // there is some fuckery on the results fqe struct, cant equality check mont_repr
+    // if(results[0] == _h4_1) {
+    //   correct++;
+    // }
+
+    
+    // Print a brief summary detailing the results
+    //
+    //printf("Computed '%d/%d' correct fq3 values!\n", correct, count);
+    // Shutdown and cleanup
+    //
+    clReleaseMemObject(input_x);
+    clReleaseMemObject(input_y);
+    clReleaseMemObject(ocl_output);
+    clReleaseProgram(kern.program);
+    clReleaseKernel(kernel);
+    clReleaseCommandQueue(kern.commands);
+    clReleaseContext(kern.context);
+    // OPENCL END
+    //break;    
+  //domain->data->iFFT(data);
 }
 
 void mnt4753_libsnark::domain_cosetFFT(
